downloader: extend it to optionally download the artifacts from Azure Pipelines
In addition to determining the latest successful build, we actually may
want to download the artifact. We now return also a function to do that.

The separation between determining the ID and the actual download is
required because we will later on want to cache the artifact, in which
case we need an ID to look up whether it has already been cached, and
only download it if it was not found in the cache.

In that vein, we also no longer only return the build ID, but prefix it
to identify the actual artifact flavor (so that we would not mistake a
cached "makepkg-git" SDK for a "build-installer" one, for example.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>