Support downloading the full SDK, too
This one is quite a bit trickier: it is offered as a `.tar.xz` because
downloading from Azure Pipelines Artifacts is not the fastest thing
(unless you can use some internal functionality, which the "Download
build artifacts" task seems to be able to do, but a GitHub Action would
not have access to that functionality).

It would appear that there are no non-native Node.js packages that _can_
uncompress `.xz`, therefore we simply use the fact that Git for Windows
is installed on the GitHub Actions' build agents, and that it includes a
`tar.exe` and an `xz.exe` to do the job for us.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>