Decouple downloading and extracting the full SDK
Even offloading the HTTPS transport to `curl.exe` seems not to have been
enough to fend off those pesky `ECONNRESET` errors. Let's offload even
the unzipping to `unzip.exe`. Sadly, this means that we do not have a
streaming unzipper anymore.

It appears that we cannot even pipe the output from one spawned process
via Node to another one: in the 64-bit case, it makes the difference
between ~9m15s and ~8m, so it makes quite a difference.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>