Turn off caching for the `full` flavor by default
Some testing reveals that the 64-bit and the 32-bit `full` SDK weigh in
with ~1,300MB and ~915MB at time of writing, respectively. Packing and
saving these to the cache takes a long time: ~5m20s and ~2m45s.

Given that the non-cached download takes ~8m20s and ~6m, respectively,
and restoring the artifacts from cache takes ~8m15s and ~5m (which is
not a lot faster, not by a long stretch), it is neither worth the time
nor the disk space to cache the `full` SDKs.

Let's introduce a new default value for the input parameter `cache`,
`auto`, that turns on caching except for the `full` flavor.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>