Retry when querying Azure Pipelines' artifacts
We encountered a problem where Azure Pipelines would return a 503. This
is obviously a transient error, and we want to be gentle and nice about
it and just retry a couple times, with a nice and easy exponential
back-off.

Sadly, `node-fetch` does not provide that functionality out of the box.
But there is a small wrapper around it that does. We just need to add
Typescript declarations, and use it, and we're all set.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>