HACK: drop invalid cached artifact
Previously, this Action did not detect when the `.zip` file was
downloaded incompletely. This actually happened, and the test fails
because of it: there are files missing, resulting in unexpected output.

A PR was merged in the meantime that should address this by detecting
the failed download and trying again.

However, the cached artifact still is there, and still causes problems.

Let's ignore the known-bad cached artifact, as identified by the key
`git-sdk-64-minimal-71789`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>