Offer to download the `vcpkg` artifacts, too
This Action essentially downloads Build Artifacts from the latest
successful run of a given Azure Pipeline.

That is exactly what we need to do with the `vcpkg` artifacts, too.

So while these artifacts are not exactly covered by the name of the
Action (the `vcpkg` artifacts have nothing to do with Git for Windows'
SDK, after all), it is easy to do.

This will help working around the issue where `actions/cache@v2.1.3` had
a bug that would prevent it to work with slashes in the path on Windows,
a bug that was fixed in v2.1.4, but `actions/cache@v2` was forced back
to v2.1.3 because of a regression (on macOS, this time) that was
introduced by v2.1.4. See https://github.com/actions/cache/issues/528
for details.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>