Drop no longer necessary hack
In d92e1c2 (HACK: drop invalid cached artifact, 2021-03-06), we added a
hack to work around a particular, incorrectly cached `minimal` artifact.

In the meantime, the bug was fixed, and a newer build is available,
therefore that hack is no longer necessary.

As it is no longer needed, let's drop it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>