ci: do not expect the source map to be stable
Apparently there is some sort of instability regarding the source maps.
In this developer's setup, a sequence 'AACA;AAIA;' was replaced by
'AACA,2BAGA;AACA;' relative to the CI setup.

The purpose of those mappings is to figure out from stacktraces where
the culprit lies in the actual original code _in Typescript_. It is
unclear where those differences come from, but then, they are not a
blocker at all for running this GitHub Action. So let's ignore them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>