"Downgrade" to node.js v12.x
GitHub Actions run on node.js v12.x, therefore we cannot use any newer
features. This lead to an error in a recent upgrade when trying to
unpack and download the full SDK:

	TypeError: fs_1.default.promises.rm is not a function

By fixing @types/node to a version compatible with node.js v12.x we will
avoid this issue in the future.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>