ci(package): allow workflow to access the environment secrets
When Dependabot pushes the branch, the workflow would run from the tip
of the pushed branch if triggered via the `push` event, hence it would
be unsafe to give it access to the environment secrets.

Instead, let's trigger on `pull_request_target`, which will run the
workflow as per `main`, where it is safe to access the environment
secrets.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>