npm-run-package.yml: avoid environment permission problems
Apparently the `if:` condition is not checked before the job is
deployed to the environment. Therefore, if the environment is protected
(because secrets!), the run fails, even if nothing was supposed to run.

Let's split it into two jobs, so that the first one can determine
whether the environment is actually needed, and skip the entire problem
if it isn't.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>