Avoid using Pipeline Artifacts, except for the `minimal` flavor
We ran into trouble with the `build-installers` flavor, where the Action
will try to download the Azure Pipeline Artifact, get only an incomplete
file, try again, and again, until the workflow run times out, see
https://github.com/git-for-windows/setup-git-for-windows-sdk/issues/336

This presented an excellent opportunity to re-think the entire strategy.

Recap: so far, we use an Azure Pipeline that is triggered whenever
anything was pushed to the Git for Windows SDK's mirror repositories
(i.e. Git repositories capturing snapshots of Git for Windows' MSYS2
fork at https://github.com/git-for-windows/git-sdk-64 and .../git-sdk-32
for x86_64 and i686, respectively). This builds the `makepkg-git`,
`build-installers` and `full` artifacts, representing various subsets of
the files. This here `setup-git-for-windows-sdk` Action, then, finds the
latest successful run and downloads the respective build artifact from
there.

Building these artifacts is not cheap, and downloading them isn't,
either, because a `.zip` file is constructed on the fly every time this
happens. For this reason, the `setup-git-for-windows-sdk` tries to cache
the artifact using GitHub Actions' native @actions/cache mechanism.

When this scheme was concocted, partial clones were not really a thing,
and GitHub's server side struggled with serving large repositories,
which is the reason why Git for Windows historically leaned a bit on
Azure Repos (which historically support large repositories better).

Fast-forward to 2022, where partial clone _is_ a thing, and performance
of shallow and partial clones is quite nice, and parallel checkout also
exists, and all of a sudden the original design decisions start making a
lot less sense.

In a semi-scientific test, then, a different design was tested: Instead
of relying on Azure Pipeline artifacts, a bare clone of the respective
SDK repository was downloaded, then the `build-extra` repository was
cloned, and the SDK artifacts were constructed from scratch. The clones
were all shallow, the SDK clone even bare, and to weigh our best
options, both partial and full clones of the SDK repositories were used.

The timings of one such run can be seen at
https://github.com/git-for-windows/build-extra/actions/runs/2475342609:

			partial		full
minimal			   48s		2m 08s
makepkg-git		1m 42s		2m 40s
build-installers	4m 35s		4m 52s
full			6m 8s		3m 02s

build-installers(i686)	2m 56s		3m 28s
full(i686)		3m 44s		2m 03s

Note: the non-i686 versions include the i686 versions, which is the
reason why the i686 versions look so much quicker.

These numbers strongly suggest to use the full clones for the full
flavor, and the partial clones for all other flavors.

These numbers _also_ suggest that the performance of this strategy is
pretty much on-par or even better than the current strategy. At
https://dev.azure.com/git-for-windows/git/_build?definitionId=34 we can
see that the latest Git artifacts Pipeline (which does something
equivalent to the current strategy to download the `build-installers`
artifact, but does not loop until it times out) takes about seven
minutes. The 2m19s illustrated above beat that hands-down.

So let's switch strategies to make everything quicker: use partial/full,
shallow, bare clones of the SDK, and a shallow clone of
https://github.com/git-for-windows/build-extra, and then run the script
that generates the SDK artifacts.

Note: we do _not_ do the same for the `minimal` flavor. The reason is
that there is more to the `git-sdk-64-minimal` Pipeline at
https://dev.azure.com/Git-for-Windows/git/_build?definitionId=22 than
just building that artifact: Since this artifact is used in all Git for
Windows _and also_ all Git CI/PR builds, we *must* avoid introducing
breakages due to changes in MSYS2 (this happened a couple of times in
the past). To avoid that, this Pipeline not only builds the artifact but
then also verifies that it can build a current Git revision and passes
the test suite. Therefore, we leave the `minimal` flavor alone, still
preferring to download that Pipeline artifact over the (otherwise quite
quick!) partial clone/sparse checkout strategy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>