Use a ridiculous number of workers for parallel checkout
It would appear that even on a 2-core build agent, using 64 workers
leads to much faster checkout times than using a single worker, at least
on the Windows agents, which is what we're interested in here.

At some stage (somewhere between 64 and 92 workers in my tests), we risk
using too many resources, though, and the checkout fails. _Very fast_,
but it fails.

So let's try to strike a balance between getting fast and still staying
robust enough not to fail.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>