Avoid picking up a `git.exe` from the current directory
When running Git's test suite, there is a `git.exe` in the current
directory, and trying to use it fails because it won't look in the
correct place for `git-remote-https.exe`. This is the symptom:

	⏷ Cloning git-sdk-64
	  Notice: Cloning https://github.com/git-for-windows/git-sdk-64 to .tmp
	  git: 'remote-https' is not a git command. See 'git --help'.

Let's fix this bug by always specifying the full path to the Git
executable we intend to be used.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>