Stop relying on Git for Windows being installed
This Action previously relied on Git for Windows to be installed in
`C:\Program Files\Git`. That is true for all hosted Windows agents.

However, this is not necessarily true for all self-hosted runners.

But there usually _is_ a Git for Windows we can rely on: The build
agent's own copy. Let's use that instead.

This fixes https://github.com/git-for-windows/setup-git-for-windows-sdk/issues/472

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>