git: export useful functionality
In the upcoming commit, we will implement support for removing the SDK
files in a post-action step. To this end, we need access to Git for
Windows' `bash` (if it exists on the agent) and to the default location
of the SDK that is written by this Action. Let's export those to allow
for reusing the functionality in other source files.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>