Avoid cluttering the workflow runs' output
When reporting anything using `core.notice()`, it will be added to the
workflow runs' summary page. This can be problematic in workflows that
use this here Action multiple times, e.g. in matrix builds. For example,
Git's own CI run for v2.38.0 reported the same commit no less than
twenty-two times: https://github.com/git/git/actions/runs/3175982844

This is excessive, and has the potential of hiding relevant, actionable
notifications among a lot of clutter.

Let's downgrade it to a mere info that is logged in the console, i.e. in
the respective job's own log, but keep it away from the workflow runs'
summary page.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>