Work around a problem with a GCC upgrade in MSYS2
The MSYS2 project upgraded their GCC from v13.2 to v14.1, and that broke
Git's CI builds on Windows when building with `DEVELOPER=1` by virtue of
Git's CI builds using the `setup-git-for-windows-sdk` GitHub Action to
set up a minimal subset of Git for Windows' SDK to build Git, and this
GitHub Action picking up that GCC upgrade by mistake.

The mistake: The `ci-artifacts` workflow of `git-sdk-64` (which is
used as an indicator whether the revision can be used via the minimal
Git for Windows SDK in Git's CI builds) did _not_ test-build with
`DEVELOPER=1` until https://github.com/git-for-windows/git-sdk-64/pull/83
was merged.

Let's specifically exclude the known-bad `git-sdk-64` revision (for
which the `ci-artifacts` workflow succeeded even when it should not
have) from being used in `setup-git-for-windows-sdk`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>