Avoid faulty cached version
I just fixed a bug where the `minimal-sdk` was already using the latest
`git-sdk-64` revision, which is not necessarily the same as the desired
`head_sha` indicates.

However, the cache ID includes that `head_sha`, which means that all
cached artifacts are bogus that use a cache ID indicating desired
`git-sdk-64` revision. So let's ignore them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>