Fix the `PATH` with which `please.sh` is invoked
The `please.sh` script is supposed to be invoked with the `PATH` that
contains a working `git.exe`. However, the way it is currently invoked,
only the `/usr/bin/` directory is added to the `PATH`, but not the
clang/MINGW directory that contains the native `git.exe`.

This does not matter on GitHub-hosted runners because Git for Windows is
installed on those runners and therefore `git.exe` is _already_ in the
`PATH`.

However, on self-hosted runners, Git for Windows may not even be
installed, and even if it is, it may not have been added to the `PATH`.

Therefore, let's add those clang/MINGW directories.

Do not even bother testing whether those directories exist; Those will
simply be ignored anyway, and it is the simplest way to guarantee that
`please.sh` will find the intended `git.exe`.

This fixes https://github.com/git-for-windows/setup-git-for-windows-sdk/issues/951.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>