minimal: get from ci-artifacts
As a result of recent changes, the Git SDK `ci-artifacts` are published
as GitHub release assets now, including several variants of the minimal
flavor of Git for Windows' SDK. This eliminates the need for us to clone
the minimal SDK in this GitHub Action.

Let's simply download the latest `.tar.gz` from the `ci-artifacts` and
extract it.

Note: As per the analysis in
https://github.com/git-for-windows/git-sdk-64/commit/fdb0cea37389, we
use the native `tar.exe` to unpack the minimal SDK while it is
downloaded, for maximal speed.

The analysis also suggests to use the `.zip` file, as this results in
the fastest operation when download speeds are above 6MB/second (which
we hope will be reliably the case in GitHub Actions). However, we want
to pipe the archive to `tar -xf -` while fetching, and it seems that for
some reason `C:\Windows\system32\tar.exe` misses `.sparse/` and `etc/`
when extracting `.zip` files from `stdin`, but the same is not true with
`.tar.gz` files. So let's use the latter.

See also: https://github.com/git-for-windows/git-sdk-64/pull/87.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>