getViaCIArtifacts: add some resilience
Due to the fact that GitHub Release assets cannot be overwritten
atomically, there is an off-chance that the asset we're looking for is
not actually there.

In those rare circumstances when we try to download such an asset while
it has been deleted so that a new version can be uploaded, just wait a
little and try again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>