azure-pipelines-task: add an Azure Pipelines task
Introduce a new sub-package that is an Azure Pipelines task, that wraps
the logic from the GitHub action.

Most of the @actions/core functions have 1:1 equivalents in Azure
Pipelines, so we can easily port the action logic into an Azure task
instead.

@actions/cache doesn't have the same integration, but consumers can use
the Cache@2 task directly in their pipeline YAML instead.

Note that due to difference in the task/action definition schema, some
of the input names cannot be exactly the same. Namely the `github-token`
action input is not a valid task input, due to the '-'. We maintain a
mapping of GitHub Actions -> Azure Pipelines input names and do the
translation in our ICore implemention.

Running `npm run build` followed by `npm run package` will use `ncc` to
compile all the task code into a single index.js file, and then use the
`tfx` CLI to create a VSIX extension that can be published to the Azure
DevOps Marketplace.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>